[fcos] Migrate to Ignition Config Spec v3.0.0
Include an Ignition fork with preliminary HTTPHeaders support,
awaiting upstream addition.

Replace "encoding/json" with dropin replacement "github.com/clarketm/json"
that supports zero values of structs with omittempty when marshaling.
See [1] for context.

[1]  https://github.com/golang/go/issues/11939

+++ pkg/daemon/update.go
@@ -3,7 +3,6 @@ package daemon
 import (
 	"bufio"
 	"bytes"
-	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"os"
@@ -17,8 +16,9 @@ import (
 	"syscall"
 	"time"
 
-	ign "github.com/coreos/ignition/config/v2_2"
-	igntypes "github.com/coreos/ignition/config/v2_2/types"
+	"github.com/clarketm/json"
+	ignConfigV3 "github.com/coreos/ignition/v2/config/v3_0"
+	ignTypes "github.com/coreos/ignition/v2/config/v3_0/types"
 	"github.com/golang/glog"
 	"github.com/google/renameio"
 	mcfgv1 "github.com/openshift/machine-config-operator/pkg/apis/machineconfiguration.openshift.io/v1"
@@ -180,7 +180,7 @@ func canonicalizeEmptyMC(config *mcfgv1.MachineConfig) *mcfgv1.MachineConfig {
 	if config != nil {
 		return config
 	}
-	newIgnCfg := ctrlcommon.NewIgnConfig()
+	newIgnCfg := ctrlcommon.NewIgnConfigSpecV3()
 	rawNewIgnCfg, err := json.Marshal(newIgnCfg)
 	if err != nil {
 		// This should never happen
@@ -302,11 +302,11 @@ func (dn *Daemon) update(oldConfig, newConfig *mcfgv1.MachineConfig) (retErr err
 		}
 	}()
 
-	oldIgnConfig, report, err := ign.Parse(oldConfig.Spec.Config.Raw)
+	oldIgnConfig, report, err := ignConfigV3.Parse(oldConfig.Spec.Config.Raw)
 	if err != nil {
 		return fmt.Errorf("parsing old Ignition config failed with error: %v\nReport: %v", err, report)
 	}
-	newIgnConfig, report, err := ign.Parse(newConfig.Spec.Config.Raw)
+	newIgnConfig, report, err := ignConfigV3.Parse(newConfig.Spec.Config.Raw)
 	if err != nil {
 		return fmt.Errorf("parsing new Ignition config failed with error: %v\nReport: %v", err, report)
 	}
@@ -390,11 +390,11 @@ func canonicalizeKernelType(kernelType string) string {
 
 // NewMachineConfigDiff compares two MachineConfig objects.
 func NewMachineConfigDiff(oldConfig, newConfig *mcfgv1.MachineConfig) (*MachineConfigDiff, error) {
-	oldIgn, report, err := ign.Parse(oldConfig.Spec.Config.Raw)
+	oldIgn, report, err := ignConfigV3.Parse(oldConfig.Spec.Config.Raw)
 	if err != nil {
 		return nil, fmt.Errorf("parsing old Ignition config failed with error: %v\nReport: %v", err, report)
 	}
-	newIgn, report, err := ign.Parse(newConfig.Spec.Config.Raw)
+	newIgn, report, err := ignConfigV3.Parse(newConfig.Spec.Config.Raw)
 	if err != nil {
 		return nil, fmt.Errorf("parsing new Ignition config failed with error: %v\nReport: %v", err, report)
 	}
@@ -434,27 +434,15 @@ func (d *MachineConfigDiff) IsEmpty() bool {
 // config currently.
 func Reconcilable(oldConfig, newConfig *mcfgv1.MachineConfig) (*MachineConfigDiff, error) {
 	// The parser will try to translate versions less than maxVersion to maxVersion, or output an err.
-	// The ignition output in case of success will always have maxVersion
-	oldIgn, report, err := ign.Parse(oldConfig.Spec.Config.Raw)
+	// First check if this is a generally valid Ignition config with a supported version
+	oldIgn, report, err := ignConfigV3.Parse(oldConfig.Spec.Config.Raw)
 	if err != nil {
-		return nil, fmt.Errorf("parsing old Ignition config failed with error: %v\nReport: %v", err, report)
+		return nil, fmt.Errorf("parsing old Ignition config failed: %v\nReport: %v", err, report)
 	}
-	newIgn, report, err := ign.Parse(newConfig.Spec.Config.Raw)
+
+	newIgn, report, err := ignConfigV3.Parse(newConfig.Spec.Config.Raw)
 	if err != nil {
-		return nil, fmt.Errorf("parsing new Ignition config failed with error: %v\nReport: %v", err, report)
-	}
-
-	// Check if this is a generally valid Ignition Config
-	if err := ctrlcommon.ValidateIgnition(newIgn); err != nil {
-		return nil, err
-	}
-
-	// Networkd section
-
-	// we don't currently configure the network in place. we can't fix it if
-	// something changed here.
-	if !reflect.DeepEqual(oldIgn.Networkd, newIgn.Networkd) {
-		return nil, errors.New("ignition networkd section contains changes")
+		return nil, fmt.Errorf("parsing new Ignition config failed: %v\nReport: %v", err, report)
 	}
 
 	// Passwd section
@@ -514,7 +502,7 @@ func Reconcilable(oldConfig, newConfig *mcfgv1.MachineConfig) (*MachineConfigDif
 	// Special case files append: if the new config wants us to append, then we
 	// have to force a reprovision since it's not idempotent
 	for _, f := range newIgn.Storage.Files {
-		if f.Append {
+		if len(f.Append) > 0 {
 			return nil, fmt.Errorf("ignition file %v includes append", f.Path)
 		}
 	}
@@ -543,8 +531,8 @@ func Reconcilable(oldConfig, newConfig *mcfgv1.MachineConfig) (*MachineConfigDif
 // Otherwise, an error will be returned and the proposed config will not be reconcilable.
 // At this time we do not support non-"core" users or any changes to the "core" user
 // outside of SSHAuthorizedKeys.
-func verifyUserFields(pwdUser igntypes.PasswdUser) error {
-	emptyUser := igntypes.PasswdUser{}
+func verifyUserFields(pwdUser ignTypes.PasswdUser) error {
+	emptyUser := ignTypes.PasswdUser{}
 	tempUser := pwdUser
 	if tempUser.Name == coreUserName && len(tempUser.SSHAuthorizedKeys) >= 1 {
 		tempUser.Name = ""
@@ -780,11 +768,11 @@ func (dn *Daemon) switchKernel(oldConfig, newConfig *mcfgv1.MachineConfig) error
 // touched.
 func (dn *Daemon) updateFiles(oldConfig, newConfig *mcfgv1.MachineConfig) error {
 	glog.Info("Updating files")
-	oldIgnConfig, report, err := ign.Parse(oldConfig.Spec.Config.Raw)
+	oldIgnConfig, report, err := ignConfigV3.Parse(oldConfig.Spec.Config.Raw)
 	if err != nil {
 		return fmt.Errorf("failed to update files. Parsing old Ignition config failed with error: %v\nReport: %v", err, report)
 	}
-	newIgnConfig, report, err := ign.Parse(newConfig.Spec.Config.Raw)
+	newIgnConfig, report, err := ignConfigV3.Parse(newConfig.Spec.Config.Raw)
 	if err != nil {
 		return fmt.Errorf("failed to update files. Parsing new Ignition config failed with error: %v\nReport: %v", err, report)
 	}
@@ -804,7 +792,7 @@ func (dn *Daemon) updateFiles(oldConfig, newConfig *mcfgv1.MachineConfig) error 
 // all the files, units that are present in the old config but not in the new one.
 // this function will error out if it fails to delete a file (with the exception
 // of simply warning if the error is ENOENT since that's the desired state).
-func (dn *Daemon) deleteStaleData(oldIgnConfig, newIgnConfig *igntypes.Config) error {
+func (dn *Daemon) deleteStaleData(oldIgnConfig, newIgnConfig *ignTypes.Config) error {
 	glog.Info("Deleting stale data")
 	newFileSet := make(map[string]struct{})
 	for _, f := range newIgnConfig.Storage.Files {
@@ -882,7 +870,7 @@ func (dn *Daemon) deleteStaleData(oldIgnConfig, newIgnConfig *igntypes.Config) e
 }
 
 // enableUnit enables a systemd unit via symlink
-func (dn *Daemon) enableUnit(unit igntypes.Unit) error {
+func (dn *Daemon) enableUnit(unit ignTypes.Unit) error {
 	// The link location
 	wantsPath := filepath.Join(wantsPathSystemd, unit.Name)
 	// sanity check that we don't return an error when the link already exists
@@ -902,7 +890,7 @@ func (dn *Daemon) enableUnit(unit igntypes.Unit) error {
 }
 
 // disableUnit disables a systemd unit via symlink removal
-func (dn *Daemon) disableUnit(unit igntypes.Unit) error {
+func (dn *Daemon) disableUnit(unit ignTypes.Unit) error {
 	// The link location
 	wantsPath := filepath.Join(wantsPathSystemd, unit.Name)
 	// sanity check so we don't return an error when the unit was already disabled
@@ -916,13 +904,13 @@ func (dn *Daemon) disableUnit(unit igntypes.Unit) error {
 }
 
 // writeUnits writes the systemd units to disk
-func (dn *Daemon) writeUnits(units []igntypes.Unit) error {
+func (dn *Daemon) writeUnits(units []ignTypes.Unit) error {
 	for _, u := range units {
 		// write the dropin to disk
 		for i := range u.Dropins {
 			glog.Infof("Writing systemd unit dropin %q", u.Dropins[i].Name)
 			dpath := filepath.Join(pathSystemd, u.Name+".d", u.Dropins[i].Name)
-			if err := writeFileAtomicallyWithDefaults(dpath, []byte(u.Dropins[i].Contents)); err != nil {
+			if err := writeFileAtomicallyWithDefaults(dpath, []byte(*u.Dropins[i].Contents)); err != nil {
 				return fmt.Errorf("failed to write systemd unit dropin %q: %v", u.Dropins[i].Name, err)
 			}
 
@@ -933,7 +921,7 @@ func (dn *Daemon) writeUnits(units []igntypes.Unit) error {
 
 		// check if the unit is masked. if it is, we write a symlink to
 		// /dev/null and continue
-		if u.Mask {
+		if u.Mask != nil && *u.Mask {
 			glog.V(2).Info("Systemd unit masked")
 			if err := os.RemoveAll(fpath); err != nil {
 				return fmt.Errorf("failed to remove unit %q: %v", u.Name, err)
@@ -948,11 +936,11 @@ func (dn *Daemon) writeUnits(units []igntypes.Unit) error {
 			continue
 		}
 
-		if u.Contents != "" {
+		if u.Contents != nil && *u.Contents != "" {
 			glog.Infof("Writing systemd unit %q", u.Name)
 
 			// write the unit to disk
-			if err := writeFileAtomicallyWithDefaults(fpath, []byte(u.Contents)); err != nil {
+			if err := writeFileAtomicallyWithDefaults(fpath, []byte(*u.Contents)); err != nil {
 				return fmt.Errorf("failed to write systemd unit %q: %v", u.Name, err)
 			}
 
@@ -965,14 +953,6 @@ func (dn *Daemon) writeUnits(units []igntypes.Unit) error {
 		// otherwise the unit is disabled. run disableUnit to ensure the unit is
 		// disabled. even if the unit wasn't previously enabled the result will
 		// be fine as disableUnit is idempotent.
-		// Note: we have to check for legacy unit.Enable and honor it
-		glog.Infof("Enabling systemd unit %q", u.Name)
-		if u.Enable {
-			if err := dn.enableUnit(u); err != nil {
-				return err
-			}
-			glog.V(2).Infof("Enabled systemd unit %q", u.Name)
-		}
 		if u.Enabled != nil {
 			if *u.Enabled {
 				if err := dn.enableUnit(u); err != nil {
@@ -992,11 +972,11 @@ func (dn *Daemon) writeUnits(units []igntypes.Unit) error {
 
 // writeFiles writes the given files to disk.
 // it doesn't fetch remote files and expects a flattened config file.
-func (dn *Daemon) writeFiles(files []igntypes.File) error {
+func (dn *Daemon) writeFiles(files []ignTypes.File) error {
 	for _, file := range files {
 		glog.Infof("Writing file %q", file.Path)
 
-		contents, err := dataurl.DecodeString(file.Contents.Source)
+		contents, err := dataurl.DecodeString(*file.Contents.Source)
 		if err != nil {
 			return err
 		}
@@ -1004,15 +984,10 @@ func (dn *Daemon) writeFiles(files []igntypes.File) error {
 		if file.Mode != nil {
 			mode = os.FileMode(*file.Mode)
 		}
-		var (
-			uid, gid = -1, -1
-		)
 		// set chown if file information is provided
-		if file.User != nil || file.Group != nil {
-			uid, gid, err = getFileOwnership(file)
-			if err != nil {
-				return fmt.Errorf("failed to retrieve file ownership for file %q: %v", file.Path, err)
-			}
+		uid, gid, err := getFileOwnership(file)
+		if err != nil {
+			return fmt.Errorf("failed to retrieve file ownership for file %q: %v", file.Path, err)
 		}
 		if err := createOrigFile(file.Path); err != nil {
 			return err
@@ -1052,31 +1027,27 @@ func createOrigFile(fpath string) error {
 }
 
 // This is essentially ResolveNodeUidAndGid() from Ignition; XXX should dedupe
-func getFileOwnership(file igntypes.File) (int, int, error) {
+func getFileOwnership(file ignTypes.File) (int, int, error) {
 	uid, gid := 0, 0 // default to root
-	if file.User != nil {
-		if file.User.ID != nil {
-			uid = *file.User.ID
-		} else if file.User.Name != "" {
-			osUser, err := user.Lookup(file.User.Name)
-			if err != nil {
-				return uid, gid, fmt.Errorf("failed to retrieve UserID for username: %s", file.User.Name)
-			}
-			glog.V(2).Infof("Retrieved UserId: %s for username: %s", osUser.Uid, file.User.Name)
-			uid, _ = strconv.Atoi(osUser.Uid)
+	if file.User.ID != nil {
+		uid = *file.User.ID
+	} else if file.User.Name != nil && *file.User.Name != "" {
+		osUser, err := user.Lookup(*file.User.Name)
+		if err != nil {
+			return uid, gid, fmt.Errorf("failed to retrieve UserID for username: %v", file.User.Name)
 		}
+		glog.V(2).Infof("Retrieved UserId: %s for username: %s", osUser.Uid, *file.User.Name)
+		uid, _ = strconv.Atoi(osUser.Uid)
 	}
-	if file.Group != nil {
-		if file.Group.ID != nil {
-			gid = *file.Group.ID
-		} else if file.Group.Name != "" {
-			osGroup, err := user.LookupGroup(file.Group.Name)
-			if err != nil {
-				return uid, gid, fmt.Errorf("failed to retrieve GroupID for group: %s", file.Group.Name)
-			}
-			glog.V(2).Infof("Retrieved GroupID: %s for group: %s", osGroup.Gid, file.Group.Name)
-			gid, _ = strconv.Atoi(osGroup.Gid)
+	if file.Group.ID != nil {
+		gid = *file.Group.ID
+	} else if file.Group.Name != nil && *file.Group.Name != "" {
+		osGroup, err := user.LookupGroup(*file.Group.Name)
+		if err != nil {
+			return uid, gid, fmt.Errorf("failed to retrieve GroupID for group: %v", file.Group.Name)
 		}
+		glog.V(2).Infof("Retrieved GroupID: %s for group: %s", osGroup.Gid, *file.Group.Name)
+		gid, _ = strconv.Atoi(osGroup.Gid)
 	}
 	return uid, gid, nil
 }
@@ -1098,7 +1069,7 @@ func (dn *Daemon) atomicallyWriteSSHKey(keys string) error {
 }
 
 // Update a given PasswdUser's SSHKey
-func (dn *Daemon) updateSSHKeys(newUsers []igntypes.PasswdUser) error {
+func (dn *Daemon) updateSSHKeys(newUsers []ignTypes.PasswdUser) error {
 	if len(newUsers) == 0 {
 		return nil
 	}
